# schnorr: Optimize sig verify with specialized types.