**This requires #2150**.

This modifies the signature verification function use the `ModNScalar` and `FieldVal` types instead of big ints.

This is work towards eventually using the specialized types throughout.

The following benchmark shows a before and after comparison of typical signature verification:

```
benchmark            old ns/op    new ns/op    delta
------------------------------------------------------
BenchmarkSigVerify   203206       182142       -10.37%

benchmark            old allocs   new allocs   delta
------------------------------------------------------
BenchmarkSigVerify   45           21           -53.33%

benchmark            old bytes    new bytes    delta
------------------------------------------------------
BenchmarkSigVerify   2066         1041         -49.61%
```